feat: Add `zoom`, `exposure` and `torchMode` to `<SkiaCamera>`
